Formatting, vectorization, Numba
Similar to my previous commit, vectorization was applied where possible and Numba annotations were used. The wherevec method breaks for some reason when annotated, so some performance is lost on this and I don't even want to think on how complicated it would be for get_arr_gra working with Numba. Better performance would probably require a rewrite, which I would rather do in something like Julia or Fortran.