Numba-compatible function
Since Numba does not support np.all, and a compiled function is faster due to additional optimizations when compared to my previous solution, we are back to looping over the rows here.

The get_arr_graph function is much harder to get working with Numba, so I will think about this more carefully before attempting to do so.