Adding Cython function
This particular function is called extensively, so any
improvements lead to great speed ups. Numba was already good, but
the Cython implementation is even better. This does make me
consider implementing more code in Cython, but this will require
careful profiling so that I can identify actual bottlenecks. For
the time being, I will test the geration of SPNs with larger
reachability graphs to stress the functions.